+++ test/libsolidity/syntaxTests/immutable/decrement.sol
@@ -1,8 +1,8 @@
 contract C {
-    uint immutable x = 3;
+    uint immutable x;
     constructor() {
         x--;
     }
 }
 // ----
-// TypeError 7733: (67-68): Immutable variables cannot be read during contract creation time, which means they cannot be read in the constructor or any function or modifier called from it.
+// TypeError 7733: (63-64): Immutable variables cannot be read during contract creation time, which means they cannot be read in the constructor or any function or modifier called from it.